refactor: 주석 삭제, 파일명 변경 등코드리뷰 반영
