+++ libs/utils-frontend/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/asteriskify';
-export * from './lib/handleCookie';
+export * from './lib/cookieManager';
 export * from './lib/checkGoodsPurchasable';
 export * from './lib/createChakraTheme';
 export * from './lib/createMuiTheme';
